fix: move test multiple-proposer-blocks into test-e2e-protocol
