+++ package.json
@@ -9,7 +9,6 @@
     "test-e2e-protocol": "LOG_LEVEL=error mocha --timeout 0 --bail --parallel --exit test/e2e/protocol/*.test.mjs ",
     "test-gas": "mocha --timeout 0 --bail --exit test/e2e/gas.test.mjs ",
     "test-e2e-tokens": "LOG_LEVEL=error mocha --timeout 0 --bail --exit test/e2e/tokens/*.test.mjs ",
-    "test-multiple-proposed-blocks": "mocha --timeout 0 --bail --exit test/e2e/multiple-proposed-blocks.test.mjs ",
     "lint": "eslint . --ext js,mjs,jsx && find-unused-exports",
     "prepare": "husky install",
     "doc:build:sdk": "jsdoc -c jsdoc.json cli/lib/nf3.mjs",
